Client : Refactor useShowInstitutionColumn.ts. Repositioned
